move code editor/pdf preview into their own folder
